Implement let bindings and variables
This does not yet create optimal BEAM ASM. In particular, each new binding
results in a new X register being used, so the output will contain lots of
sequences that look like

{move,{x,0},{x,1}}.
{move,{x,1},{x,0}}.

...for bindings that will be used immediately in the next binding or the nested
expression.

I want to get to the next exercise to understand whether this will be necessary
before implementing ANF, but BEAM ASM relies a lot on "result registers" for
BIFs, which implies we might be able to thread through a "destination register"
to optimize these; that is, when the input and destination register are the
same, we don't need to use another. For example, in the expression `let y = 5 in
y`, the ultimate result should be put in X0 and there's only one binding, so we
don't need X1 at all.

Furthermore, BEAM uses the Y registers (on the stack, basically) for local
variables that are not used in call arguments, but you have to allocate them
before use and deallocate them when finished. This usually looks like:

{allocate, N, M}.
%% use the Y registers in some computation
{deallocate, M}.

Without knowing how deep the bindings go from the outside of the expression, we
can't preallocate them.